Move menu to components package
