+++ src/index.ts
@@ -14,8 +14,8 @@ import defaultToolbar from './layouts/defaultToolbar';
 import { default as ISlot } from './layouts/Slot';
 import { default as IToolbarSlot, RenderToolbar as RenderToolbarType, RenderToolbarSlot as RenderToolbarSlotType } from './layouts/ToolbarSlot';
 import { default as ILocalizationMap } from './localization/LocalizationMap';
-import MenuDivider from './menu/MenuDivider';
-import MenuItem from './menu/MenuItem';
+import MenuDivider from './components/MenuDivider';
+import MenuItem from './components/MenuItem';
 import Modal from './portal/Modal';
 import Popover from './portal/Popover';
 import Position from './portal/Position';
